Converted to flask app factory pattern and split routes into separate view files using Blueprint pattern.
