+++ server/models.py
@@ -1,28 +1,37 @@
-from app import db
+from flask_sqlalchemy import *
+from sqlalchemy import *
+from sqlalchemy.dialects.postgresql import ARRAY
+from sqlalchemy.ext.declarative import DeferredReflection
+from sqlalchemy.orm import relationship
 from werkzeug.security import generate_password_hash, check_password_hash
 import datetime
 
-order_product_association_table = db.Table(
+class Reflected(DeferredReflection):
+    __abstract__ = True
+
+Base = declarative_base()
+
+order_product_association_table = Table(
 	"order_product_association",
-	db.metadata,
-	db.Column("order_id", db.ForeignKey("order.id")),
-	db.Column("product_id", db.ForeignKey("product.id"))
+	Base.metadata,
+	Column("order_id", ForeignKey("order.id", ondelete="CASCADE")),
+	Column("product_id", ForeignKey("product.id", ondelete="CASCADE"))
 )
 
-class User(db.Model):
+class User(Reflected, Base):
     __tablename__ = 'user'
 
-    id = db.Column(db.Integer, primary_key=True)
-    email = db.Column(db.Text)
-    password = db.Column(db.Text)
-    mobile_number = db.Column(db.Text)
-    session_token = db.Column(db.Text)
-    sms_code = db.Column(db.Text)
-    first_name = db.Column(db.Text)
-    last_name = db.Column(db.Text)
-    address = db.Column(db.Text)
-    orders = db.relationship("Order", backref='user', lazy='dynamic')
-    active = db.Column(db.Boolean)
+    id = Column(Integer, primary_key=True)
+    email = Column(Text)
+    password = Column(Text)
+    mobile_number = Column(Text)
+    session_token = Column(Text)
+    sms_code = Column(Text)
+    first_name = Column(Text)
+    last_name = Column(Text)
+    address = Column(Text)
+    orders = relationship("Order", backref='user', lazy='dynamic', cascade="all,delete")
+    active = Column(Boolean)
 
     def __init__(self, first_name: str, last_name: str, email: str, password: str, mobile_number: str, address: str, orders: list = []):
         self.first_name = first_name
@@ -61,19 +70,19 @@ class User(db.Model):
         print(val)
         return val
 
-class Product(db.Model):
+class Product(Reflected, Base):
     __tablename__ = 'product'
 
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.Text)
-    price = db.Column(db.Integer)
-    discount = db.Column(db.Float)
-    categories = db.Column(db.ARRAY(db.Text))
-    amount_sold = db.Column(db.Integer)
-    description = db.Column(db.Text)
-    img = db.Column(db.Text)
-    inventory = db.Column(db.Integer)
-    seller = db.Column(db.Text)
+    id = Column(Integer, primary_key=True)
+    name = Column(Text)
+    price = Column(Integer)
+    discount = Column(Float)
+    categories = Column(ARRAY(Text))
+    amount_sold = Column(Integer)
+    description = Column(Text)
+    img = Column(Text)
+    inventory = Column(Integer)
+    seller = Column(Text)
 
     def __init__(self, name: str, price: int, discount: float, categories: [str], amount_sold: int, description: str,
                  img: str, inventory: int, seller: str):
@@ -102,14 +111,14 @@ class Product(db.Model):
         }
 
 
-class Order(db.Model):
+class Order(Reflected, Base):
     __tablename__ = 'order'
 
-    id = db.Column(db.Integer, primary_key=True)
-    products = db.relationship("Product", secondary=order_product_association_table)
-    user_id = db.Column(db.Integer, db.ForeignKey("user.id"))
-    created_date = db.Column(db.DateTime, default=datetime.datetime.utcnow)
-    total = db.Column(db.Integer)
+    id = Column(Integer, primary_key=True)
+    products = relationship("Product", secondary=order_product_association_table, cascade="all,delete")
+    user_id = Column(Integer, ForeignKey("user.id", ondelete="CASCADE"))
+    created_date = Column(DateTime, default=datetime.datetime.utcnow)
+    total = Column(Integer)
 
     def __init__(self, products: list, user_id):
         self.products = products
